requirements.txt
Flask
Flask-Restful
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
Werkzeug
pytest
python-dotenv
psycopg2-binary